Drop in code
This code is copied from one of my local repositories. There are dozens
of commits and maybe a month's worth of work actually. But since I want
to make this widely more widely available, I'm pasting in this
development version as it is, but changing a few things such as module
name and import paths.

Add License