build: Upgrade golang.org/x/net
Ran go mod tidy.

Addresses dependabot#1, dependabot#2. These advisories are about package
golang.org/x/net/http2, which is under the same module as another
package used in this application, golang.org/x/net/html. However, the
only packages in this tool that uses that same module are the edam and
enex packages. So it's not really affected by the advisory.